refactor: 사이드바를 별도 컴포넌트로 분리 (SP-578)
